+++ Hydra/Serializer/CollectionNormalizer.php
@@ -14,11 +14,9 @@ namespace Dunglas\ApiBundle\Hydra\Serializer;
 use Dunglas\ApiBundle\Api\Filter\FilterInterface;
 use Dunglas\ApiBundle\Api\ResourceCollectionInterface;
 use Dunglas\ApiBundle\Api\ResourceInterface;
-use Dunglas\ApiBundle\Api\ResourceResolverTrait;
-use Dunglas\ApiBundle\Exception\InvalidArgumentException;
+use Dunglas\ApiBundle\Api\ResourceResolver;
 use Dunglas\ApiBundle\JsonLd\ContextBuilder;
 use Dunglas\ApiBundle\JsonLd\Serializer\ContextTrait;
-use Dunglas\ApiBundle\Model\PaginatorInterface;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
 use Symfony\Component\Serializer\Normalizer\SerializerAwareNormalizer;
 
@@ -32,31 +30,29 @@ class CollectionNormalizer extends SerializerAwareNormalizer implements Normaliz
 {
     const FORMAT = 'jsonld';
 
-    use ResourceResolverTrait;
     use ContextTrait;
 
     /**
      * @var string
      */
     const HYDRA_COLLECTION = 'hydra:Collection';
-    /**
-     * @var string
-     */
-    const HYDRA_PAGED_COLLECTION = 'hydra:PagedCollection';
 
     /**
      * @var ContextBuilder
      */
     private $contextBuilder;
+    /**
+     * @var ResourceResolver
+     */
+    private $resourceResolver;
 
     /**
-     * @param ResourceCollectionInterface $resourceCollection
-     * @param ContextBuilder              $contextBuilder
+     * @param ContextBuilder $contextBuilder
      */
-    public function __construct(ResourceCollectionInterface $resourceCollection, ContextBuilder $contextBuilder)
+    public function __construct(ContextBuilder $contextBuilder, ResourceResolver $resourceResolver)
     {
-        $this->resourceCollection = $resourceCollection;
         $this->contextBuilder = $contextBuilder;
+        $this->resourceResolver = $resourceResolver;
     }
 
     /**
@@ -72,7 +68,7 @@ class CollectionNormalizer extends SerializerAwareNormalizer implements Normaliz
      */
     public function normalize($object, $format = null, array $context = array())
     {
-        $resource = $this->guessResource($object, $context);
+        $resource = $this->resourceResolver->guessResource($object, $context);
 
         if (isset($context['jsonld_sub_level'])) {
             $data = [];
@@ -81,137 +77,18 @@ class CollectionNormalizer extends SerializerAwareNormalizer implements Normaliz
             }
         } else {
             $context = $this->createContext($resource, $context);
-
             $data = [
                 '@context' => $this->contextBuilder->getResourceContext($resource, $context),
                 '@id' => $context['request_uri'],
+                '@type' => self::HYDRA_COLLECTION,
+                'hydra:member' => []
             ];
-            list($parts, $parameters) = $this->parseRequestUri($resource, $context['request_uri']);
 
-            if ($object instanceof PaginatorInterface) {
-                $data['@type'] = self::HYDRA_PAGED_COLLECTION;
-
-                $currentPage = $object->getCurrentPage();
-                $lastPage = $object->getLastPage();
-
-                if (1. !== $currentPage) {
-                    $previousPage = $currentPage - 1.;
-                    $data['hydra:previousPage'] = $this->getPageUrl($resource, $parts, $parameters, $previousPage);
-                }
-
-                if ($currentPage !== $lastPage) {
-                    $data['hydra:nextPage'] = $this->getPageUrl($resource, $parts, $parameters, $currentPage + 1.);
-                }
-
-                $data['hydra:totalItems'] = $object->getTotalItems();
-                $data['hydra:itemsPerPage'] = $object->getItemsPerPage();
-                $data['hydra:firstPage'] = $this->getPageUrl($resource, $parts, $parameters, 1.);
-                $data['hydra:lastPage'] = $this->getPageUrl($resource, $parts, $parameters, $lastPage);
-            } else {
-                $data['@type'] = self::HYDRA_COLLECTION;
-            }
-
-            $data['hydra:member'] = [];
             foreach ($object as $obj) {
                 $data['hydra:member'][] = $this->serializer->normalize($obj, $format, $context);
-            }
-
-            $filters = $resource->getFilters();
-            if (!empty($filters)) {
-                $data['hydra:search'] = $this->getSearch($resource, $parts, $filters);
             }
         }
 
         return $data;
-    }
-
-    /**
-     * Parse and standardize the request URI.
-     *
-     * @param ResourceInterface $resource
-     * @param string            $requestUri
-     *
-     * @return array
-     *
-     * @throws InvalidArgumentException
-     */
-    private function parseRequestUri(ResourceInterface $resource, $requestUri)
-    {
-        $parts = parse_url($requestUri);
-        if (false === $parts) {
-            throw new InvalidArgumentException(sprintf('The request URI "%s" is malformed.', $requestUri));
-        }
-
-        $parameters = [];
-        if (isset($parts['query'])) {
-            parse_str($parts['query'], $parameters);
-
-            // Remove existing page parameter
-            if (isset($parameters[$resource->getPageParameter()])) {
-                unset($parameters[$resource->getPageParameter()]);
-            }
-        }
-
-        return [$parts, $parameters];
-    }
-
-    /**
-     * Gets a collection URL for the given page.
-     *
-     * @param ResourceInterface $resource
-     * @param array             $parts
-     * @param array             $parameters
-     * @param float             $page
-     *
-     * @return string
-     */
-    private function getPageUrl(ResourceInterface $resource, array $parts, array $parameters, $page)
-    {
-        if (1. !== $page) {
-            $parameters[$resource->getPageParameter()] = $page;
-        }
-
-        $parts['query'] = http_build_query($parameters);
-
-        $url = $parts['path'];
-
-        if ('' !== $parts['query']) {
-            $url .= '?'.$parts['query'];
-        }
-
-        return $url;
-    }
-
-    /**
-     * Returns the content of the Hydra search property.
-     *
-     * @param ResourceInterface $resource
-     * @param array             $parts
-     * @param FilterInterface[] $filters
-     *
-     * @return array
-     */
-    private function getSearch(ResourceInterface $resource, array $parts, array $filters)
-    {
-        $variables = [];
-        $mapping = [];
-        foreach ($filters as $filter) {
-            foreach ($filter->getDescription($resource) as $variable => $data) {
-                $variables[] = $variable;
-                $mapping[] = [
-                    '@type' => 'IriTemplateMapping',
-                    'variable' => $variable,
-                    'property' => $data['property'],
-                    'required' => $data['required'],
-                ];
-            }
-        }
-
-        return [
-            '@type' => 'hydra:IriTemplate',
-            'hydra:template' => sprintf('%s{?%s}', $parts['path'], implode(',', $variables)),
-            'hydra:variableRepresentation' => 'BasicRepresentation',
-            'hydra:mapping' => $mapping,
-        ];
     }
 }
