+++ Resources/config/hydra.xml
@@ -5,6 +5,7 @@
            xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
 
     <services>
+
         <service id="api.hydra.documentation_builder" class="Dunglas\ApiBundle\Hydra\ApiDocumentationBuilder">
             <argument type="service" id="api.resource_collection" />
             <argument type="service" id="api.jsonld.context_builder" />
@@ -33,8 +34,8 @@
         <!-- Serializer -->
 
         <service id="api.hydra.normalizer.collection" class="Dunglas\ApiBundle\Hydra\Serializer\CollectionNormalizer" public="false">
-            <argument type="service" id="api.resource_collection" />
             <argument type="service" id="api.json_ld.context_builder" />
+            <argument type="service" id="api.resource_resolver" />
 
             <tag name="serializer.normalizer" priority="50" />
         </service>
@@ -57,6 +58,26 @@
         <service id="api.hydra.action.documentation" class="Dunglas\ApiBundle\Hydra\Action\DocumentationAction">
             <argument type="service" id="api.hydra.documentation_builder" />
         </service>
+
+        <!-- Serializer enhancers -->
+        <service
+                id="api.hydra.normalizer.paginated_collection_enhancer"
+                class="Dunglas\ApiBundle\Hydra\Serializer\PaginatedCollectionEnhancer"
+                decorates="api.hydra.normalizer.collection"
+                public="false">
+            <argument type="service" id="api.hydra.normalizer.paginated_collection_enhancer.inner" />
+            <argument type="service" id="api.resource_resolver" />
+        </service>
+
+        <service
+                id="api.hydra.normalizer.collection_filters_enhancer"
+                class="Dunglas\ApiBundle\Hydra\Serializer\CollectionFiltersEnhancer"
+                decorates="api.hydra.normalizer.collection"
+                public="false">
+            <argument type="service" id="api.hydra.normalizer.collection_filters_enhancer.inner" />
+            <argument type="service" id="api.resource_resolver" />
+        </service>
+
     </services>
 
 </container>
