+++ Resources/config/jsonld.xml
@@ -25,6 +25,7 @@
             <argument type="service" id="api.mapping.class_metadata_factory" />
             <argument type="service" id="api.jsonld.context_builder" />
             <argument type="service" id="property_accessor" />
+            <argument type="service" id="api.resource_resolver" />
             <argument type="service" id="api.name_converter" on-invalid="ignore" />
 
             <tag name="serializer.normalizer" />
